Fix/bug

+++ BE/routes/serviceNotice.js
@@ -15,92 +15,98 @@ const router = express.Router();
  * POST Service Notice Detail
  * Example URL = ../service
  *----------------------------------------------------------------------*/
-router.post(
-  "/service",
-  verifyToken,
-  service_upload.single("service_notice_image"),
-  async (req, res) => {
-    const { title, context, fixed, attachment } = req.body;
+router.post("/service", verifyToken, service_upload.single("service_img"), async (req, res) => {
+  const { hospital_id, title, context, fixed, attachment } = req.body;
+  if (hospital_id != 24) {
+    logger.info("POST Service Notice");
+    return res.json({ state: "Fail", Message: "사이트 관리자만 접근 가능합니다." });
+  }
+  const rename =
+    new Date(+new Date() + 3240 * 10000).toISOString().replace("T", " ").replace(/\..*/, "") +
+    attachment;
 
-    const rename =
-      new Date(+new Date() + 3240 * 10000).toISOString().replace("T", " ").replace(/\..*/, "") +
-      attachment;
-
-    try {
-      let a;
-      if (req.body.attachment) {
-        nameParser("uploads/service", "uploads/service", attachment, rename);
-        const sql = `INSERT INTO service_notice ( 
+  try {
+    let a;
+    if (req.body.attachment) {
+      nameParser("uploads/service", "uploads/service", attachment, rename);
+      const sql = `INSERT INTO service_notice ( 
                         title, 
                         context, 
                         fixed, 
                         attachment,
                         created_at) VALUES(?, ?, ?, ?, now());`;
-        const data = await pool.query(sql, [title, context, fixed, rename]);
-      } else {
-        const sql = `INSERT INTO service_notice ( 
+      const data = await pool.query(sql, [title, context, fixed, rename]);
+    } else {
+      const sql = `INSERT INTO service_notice ( 
                         title, 
                         context, 
                         fixed,
                         created_at) VALUES(?, ?, ?, now());`;
-        const data = await pool.query(sql, [title, context, fixed]);
-      }
-      const LAST_INSERT_ID = `SELECT MAX(id) as auto_id FROM service_notice;`;
-      const data_id = await pool.query(LAST_INSERT_ID);
-      const create_id = data_id[0][0].auto_id;
-      logger.info("POST Event Detail");
-      return res.json({ state: "Success", id: create_id });
-    } catch (error) {
-      logger.error("POST Service Notice Detail " + error);
-      return res.json({ state: "Fail" });
+      const data = await pool.query(sql, [title, context, fixed]);
     }
+    const LAST_INSERT_ID = `SELECT MAX(id) as auto_id FROM service_notice;`;
+    const data_id = await pool.query(LAST_INSERT_ID);
+    const create_id = data_id[0][0].auto_id;
+    logger.info("POST Service Notice");
+    return res.json({ state: "Success", id: create_id });
+  } catch (error) {
+    logger.error("POST Service Notice " + error);
+    return res.json({ state: "Fail" });
   }
-);
+});
 
 /*----------------------------------------------------------------------*
  * PUT Service Notice Detail
  * Example URL = ../service/1
  *----------------------------------------------------------------------*/
-router.put(
-  "/service/:id",
-  verifyToken,
-  service_upload.single("service_notice_image"),
-  async (req, res) => {
-    const id = req.params.id;
-    const { title, context, fixed, attachment } = req.body;
+router.put("/service/:id", verifyToken, service_upload.single("service_img"), async (req, res) => {
+  const id = req.params.id;
+  const { hospital_id, title, context, fixed, attachment, del } = req.body;
+  if (hospital_id != 24) {
+    logger.info("POST Service Notice");
+    return res.json({ state: "Fail", Message: "사이트 관리자만 접근 가능합니다." });
+  }
+  const rename =
+    new Date(+new Date() + 3240 * 10000).toISOString().replace("T", " ").replace(/\..*/, "") +
+    attachment;
+  // const path = "uploads/service/" + rename;
 
-    const rename =
-      new Date(+new Date() + 3240 * 10000).toISOString().replace("T", " ").replace(/\..*/, "") +
-      attachment;
-    // const path = "uploads/service/" + rename;
-
-    try {
-      if (req.body.attachment) {
-        nameParser("uploads/service", "uploads/service", attachment, rename);
-        const sql = `UPDATE service_notice SET 
+  try {
+    if (attachment) {
+      nameParser("uploads/service", "uploads/service", attachment, rename);
+      const sql = `UPDATE service_notice SET 
                       title=?, 
                       context=?, 
                       fixed=?, 
                       attachment=?,
                       updated_at = now() WHERE id=?;`;
-        const data = await pool.query(sql, [title, context, fixed, rename, id]);
-      } else {
+      const data = await pool.query(sql, [title, context, fixed, rename, id]);
+    } else {
+      if (del == 0) {
         const sql = `UPDATE service_notice SET 
                       title=?, 
                       context=?, 
                       fixed=?,
                       updated_at = now() WHERE id=?;`;
         const data = await pool.query(sql, [title, context, fixed, id]);
+      } else {
+        const sql = `UPDATE service_notice SET 
+                      title=?, 
+                      context=?, 
+                      fixed=?,
+                      attachment = null,
+                      updated_at = now() WHERE id=?;`;
+        const data = await pool.query(sql, [title, context, fixed, id]);
       }
-
-      logger.info("PUT Service Notice Detail");
-      return res.json({ result: "Success" });
-    } catch (error) {
-      logger.error("PUT Service Notice Detail " + error);
-      return res.json(error);
     }
+
+    logger.info("PUT Service Notice");
+    return res.json({ result: "Success", id: id });
+  } catch (error) {
+    logger.error("PUT Service Notice" + error);
+    return res.json(error);
   }
-);
+});
 
 /*----------------------------------------------------------------------*
  * DELETE Service Notice Detail
@@ -108,7 +114,11 @@ router.put(
  *----------------------------------------------------------------------*/
 router.delete("/service/:id", verifyToken, async (req, res) => {
   const id = req.params.id;
-
+  const { hospital_id } = req.body;
+  if (hospital_id != 24) {
+    logger.info("POST Service Notice");
+    return res.json({ state: "Fail", Message: "사이트 관리자만 접근 가능합니다." });
+  }
   try {
     const sql = `DELETE FROM service_notice WHERE id=?;`;
     const data = await pool.query(sql, [id]);
@@ -139,7 +149,8 @@ router.get("/service/:id", async (req, res) => {
     FROM service_notice WHERE ID = ?;`;
     const data = await pool.query(sql, [id, id, id, id, id]);
     let result = data[0][0];
-    result.image = "http://i6a205.p.ssafy.io:8000/api/serviceimage/" + result.attachment;
+    if (result.attachment)
+      result.image = "http://i6a205.p.ssafy.io:8000/api/serviceimage/" + result.attachment;
     logger.info("GET Service Notice Detail");
     return res.json(result);
   } catch (error) {
